+++ NEWS.md
@@ -5,6 +5,7 @@
 - Slight speed-up of `permshap()` by saving calculations for the two special permutations of all 0 and all 1.
 - Consequently, the `m_exact` component in the output is reduced by 2.
 - Slight speed-up of `permshap()` by optimizing internal function `shapley_formula()`.
+- Slight speed-up of `kernelshap()` and `permshap()` for single-output predictions.
 
 # kernelshap 0.4.0